+++ src/components/DatatableV2/body/BodyCell.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from 'react';
 import clx from 'classnames';
 
 import { getCommonCellStyles } from '../columns.utils';
-import { useHasHozintalScroll } from '../hooks/useHasHozintalScroll';
+import { useHasHorizontalScroll } from '../hooks/useHasHorizontalScroll';
 import Skeleton from './Skeleton';
 
 const BodyCell = <D,>({
@@ -20,7 +20,7 @@ const BodyCell = <D,>({
   const { isLoading } = table.getState();
 
   const [skeletonWidth, setSkeletonWidth] = useState(100);
-  const hasHorizontalScroll = useHasHozintalScroll(table);
+  const hasHorizontalScroll = useHasHorizontalScroll(table);
 
   useEffect(() => {
     if (!isLoading || skeletonWidth !== 100) return;